chore: restucture message files
